Refactor: move path, delete mood button, style
